+++ library/eos_switchport.py
@@ -107,6 +107,16 @@ options:
     choices: []
     aliases: []
     version_added: 1.0.0
+  trunk_groups:
+    description:
+      - Configures the list of trunk groups on the switchport.  The parameter
+        accepts a comma separated list of values to be provisioned on the
+        interface.
+    required: false
+    default: null
+    choices: []
+    aliases: []
+    version_added: 1.1.0
 """
 
 EXAMPLES = """
@@ -119,6 +129,9 @@ EXAMPLES = """
 
 - name: Add the set of allowed vlans to Ethernet2/1
   eos_switchport: name=Ethernet2/1 mode=trunk trunk_allowed_vlans=1,10,100
+
+- name: Add trunk group values to an interface
+  eos_switchport: name=Ethernet5 trunk_groups=foo,bar,baz
 
 """
 #<<EOS_COMMON_MODULE_START>>
@@ -178,6 +191,8 @@ class EosAnsibleModule(AnsibleModule):
         self.debug('params', self.params)
 
         self._attributes = self.map_argument_spec()
+        self.validate()
+
         self._node = self.connect()
 
         self._instance = None
@@ -236,6 +251,12 @@ class EosAnsibleModule(AnsibleModule):
         if 'CHECKMODE' in attrs:
             del attrs['CHECKMODE']
         return attrs
+
+    def validate(self):
+        for key, value in self.attributes.iteritems():
+            func = self.func('validate_%s' % key)
+            if func:
+                self.attributes[key] = func(value)
 
     def create(self):
         if not self.check_mode:
@@ -416,6 +437,7 @@ def instance(module):
         _instance['access_vlan'] = result['access_vlan']
         _instance['trunk_native_vlan'] = result['trunk_native_vlan']
         _instance['trunk_allowed_vlans'] = result['trunk_allowed_vlans']
+        _instance['trunk_groups'] = ','.join(result['trunk_groups'])
     return _instance
 
 def create(module):
@@ -468,6 +490,21 @@ def set_trunk_allowed_vlans(module):
                'with value %s' % (name, value))
     module.node.api('switchports').set_trunk_allowed_vlans(name, value)
 
+def set_trunk_groups(module):
+    """Configures the set of trunk groups on the interface
+    """
+    name = module.attributes['name']
+    value = module.attributes['trunk_groups'].split(',')
+    module.log('Invoked set_trunk_groups for eos_switchport[%s] '
+               'with value %s' % (name, value))
+    module.node.api('switchports').set_trunk_groups(name, value)
+
+def validate_trunk_groups(value):
+    """Validates the trunk_groups argument
+    """
+    values = sorted(value.split(','))
+    return ','.join(values)
+
 def main():
     """ The main module routine called when the module is run by Ansible
     """
@@ -477,7 +514,8 @@ def main():
         mode=dict(choices=['access', 'trunk']),
         access_vlan=dict(),
         trunk_native_vlan=dict(),
-        trunk_allowed_vlans=dict()
+        trunk_allowed_vlans=dict(),
+        trunk_groups=dict()
     )
 
     module = EosAnsibleModule(argument_spec=argument_spec,